CurrenciesBnbLoadingService.java added
