+++ currency-converter/src/main/java/com/example/currencyconverter/controller/AdminController.java
@@ -2,6 +2,7 @@ package com.example.currencyconverter.controller;
 
 import com.example.currencyconverter.domain.Currency;
 import com.example.currencyconverter.dto.CurrencyDto;
+import com.example.currencyconverter.service.CurrenciesBnbLoadingService;
 import com.example.currencyconverter.service.CurrencyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,13 +12,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.validation.Valid;
+import java.util.List;
 
 @Controller
 public class AdminController {
     @Autowired
     CurrencyService currencyService;
 
-
+    @Autowired
+    CurrenciesBnbLoadingService currenciesBnbLoadingService;
 
     @GetMapping("/admin/addCurrency")
     public ModelAndView addCurrency() {
@@ -36,6 +39,17 @@ public class AdminController {
                 currencyDto.getCode(), currencyDto.getName(),
                 currencyDto.getRate(), currencyDto.getRatio());
         currencyService.addCurrency(currency);
+        return "redirect:/";
+    }
+
+    @GetMapping("/admin/currencies/refresh")
+    public String refreshCurrenciesRates(){
+        List<Currency> currencyList =  currenciesBnbLoadingService.getCurrencies();
+
+        for (Currency currency : currencyList){
+            currencyService.addCurrency(currency);
+        }
+
         return "redirect:/";
     }
 }
