+++ currency-converter/src/main/resources/templates/index.html
@@ -6,9 +6,7 @@
     <meta charset="UTF-8">
     <title>Title</title>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"/>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
+
 </head>
 <body>
 <div class="container">
@@ -19,6 +17,7 @@
         <button type="button" class="btn btn-default navbar-btn" > currency</button>
         <span sec:authentication="name"></span></b>
         <a th:href="@{/admin/addCurrency}">Add a currency.</a></p>
+        <a th:href="@{/admin/currencies/refresh}">Update all currencies from BNB rates.</a></p>
         <a th:href="@{/logout}">Log out</a></p>
     </div>
     <div class="row">
@@ -81,6 +80,8 @@
     </div>
 </div>
 
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
 <script type="text/javascript" src="main.js"></script>
 </body>
 </html>
